Fix ctype compile time warnings on NetBSD (bug #908).

+++ gammu/calendar.c
@@ -942,7 +942,7 @@ void ListToDoCategory(int argc UNUSED, char *argv[])
 
 	Length = strlen(argv[2]);
 	for (j = 0; j < Length; j++) {
-		if (!isdigit(argv[2][j])) {
+		if (!isdigit((int)argv[2][j])) {
 			Number = false;
 			break;
 		}
